Almost finished ohloh.net integration, content wise.
The API is quite old and retarded, it returns only xml and
only list projects user manage, not the ones they have contributed
to.

Furthermore to get all the content I wanted, from my own account,
takes over 20 connections to ohloh servers. This is of course
unacceptably slow, and burns throught the 1000 a day allowance.